taxi-backend/requirements.txt
django
django-cors-headers
django_rest_framework
django-timescaledb
drf-yasg
psycopg2-binary
paho-mqtt
tzdata